+++ Content.Shared/Cloning/CloningPodComponent.cs
@@ -82,8 +82,12 @@ public sealed partial class CloningPodComponent : Component
 
     /// <summary>
     ///     Controls whether a cloning machine performs the Metempsychosis functions, EG: Is this a Cloner or a Metem Machine?
-    ///     Metempsychosis refers to the metaphysical process of Reincarnation
+    ///     Metempsychosis refers to the metaphysical process of Reincarnation.
     /// </summary>
+    /// <remarks>
+    ///     A Machine with this enabled will essentially create a random new character instead of creating a living version of the old character.
+    ///     Although, the specifics of how much of the new body is a "new character" is highly adjustable in server configuration.
+    /// </remarks>
     [DataField]
     public bool DoMetempsychosis;
 
@@ -141,42 +145,4 @@ public enum ForcedMetempsychosisType : byte
     Clone,
     RandomHumanoid,
     RandomNonHumanoid
-}
-
-/// <summary>
-///     Raised after a new mob got spawned when cloning a humanoid
-/// </summary>
-[ByRefEvent]
-public struct CloningEvent
-{
-    public bool NameHandled = false;
-
-    public readonly EntityUid Source;
-    public readonly EntityUid Target;
-
-    public CloningEvent(EntityUid source, EntityUid target)
-    {
-        Source = source;
-        Target = target;
-    }
-}
-
-/// <summary>
-///     Raised on a corpse being subjected to forced reincarnation(Metempsychosis). Allowing for innate effects from the mob to influence the reincarnation.
-/// </summary>
-[ByRefEvent]
-public struct ReincarnatingEvent
-{
-    public bool OverrideChance;
-    public bool NeverTrulyClone;
-    public ForcedMetempsychosisType ForcedType = ForcedMetempsychosisType.None;
-    public readonly EntityUid OldBody;
-    public float ReincarnationChanceModifier = 1;
-    public float ReincarnationChances;
-
-    public ReincarnatingEvent(EntityUid oldBody, float reincarnationChances)
-    {
-        OldBody = oldBody;
-        ReincarnationChances = reincarnationChances;
-    }
 }
